+++ tests/test_checks_other.py
@@ -7,6 +7,7 @@ from threedi_modelchecker.checks.other import ConnectionNodesLength
 from threedi_modelchecker.checks.other import CrossSectionLocationCheck
 from threedi_modelchecker.checks.other import LinestringLocationCheck
 from threedi_modelchecker.checks.other import OpenChannelsWithNestedNewton
+from threedi_modelchecker.checks.other import PotentialBreachInterdistanceCheck
 from threedi_modelchecker.checks.other import PotentialBreachStartEndCheck
 from threedi_modelchecker.checks.other import SpatialIndexCheck
 from threedi_modelchecker.threedi_model import constants
@@ -240,26 +241,66 @@ def test_spatial_index_disabled(empty_sqlite_v4):
 
 
 @pytest.mark.parametrize(
-    "breach_geom,ok",
+    "x,y,ok",
     [
-        ("LINESTRING(-71.064544 42.28787, -71.064544 42.286)", True),  # at start
-        ("LINESTRING(-71.0645 42.287, -71.0645 42.286)", True),  # at end
-        ("LINESTRING(-71.06452 42.2874, -71.0645 42.286)", True),  # middle
-        (
-            "LINESTRING(-71.064544 42.287871, -71.064544 42.286)",
-            False,
-        ),  # close to start
-        ("LINESTRING(-71.0645 42.287001, -71.0645 42.286)", False),  # close to end
+        (-71.064544, 42.28787, True),  # at start
+        (-71.0645, 42.287, True),  # at end
+        (-71.06452, 42.2874, True),  # middle
+        (-71.064544, 42.287869, False),  # close to start
+        (-71.064499, 42.287001, False),  # close to end
     ],
 )
-def test_potential_breach_start_end(session, breach_geom, ok):
+def test_potential_breach_start_end(session, x, y, ok):
     # channel geom: LINESTRING (-71.064544 42.28787, -71.0645 42.287)
-    factories.PotentialBreachFactory(the_geom=f"SRID=4326;{breach_geom}")
+    factories.PotentialBreachFactory(
+        the_geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)"
+    )
     check = PotentialBreachStartEndCheck(
-        models.PotentialBreach.the_geom, max_distance=1.0
+        models.PotentialBreach.the_geom, min_distance=1.0
     )
     invalid = check.get_invalid(session)
     if ok:
         assert len(invalid) == 0
     else:
         assert len(invalid) == 1
+
+
+@pytest.mark.parametrize(
+    "x,y,ok",
+    [
+        (-71.06452, 42.2874, True),  # exactly on other
+        (-71.06452, 42.287401, False),  # too close to other
+        (-71.0645, 42.287, True),  # far enough from other
+    ],
+)
+def test_potential_breach_interdistance(session, x, y, ok):
+    # channel geom: LINESTRING (-71.064544 42.28787, -71.0645 42.287)
+    ref = factories.PotentialBreachFactory(
+        the_geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)"
+    )
+    factories.PotentialBreachFactory(
+        the_geom=f"SRID=4326;LINESTRING({x} {y}, -71.064544 42.286)",
+        channel=ref.channel,
+    )
+    check = PotentialBreachInterdistanceCheck(
+        models.PotentialBreach.the_geom, min_distance=1.0
+    )
+    invalid = check.get_invalid(session)
+    if ok:
+        assert len(invalid) == 0
+    else:
+        assert len(invalid) == 1
+
+
+def test_potential_breach_interdistance_other_channel(session):
+    factories.PotentialBreachFactory(
+        the_geom="SRID=4326;LINESTRING(-71.06452 42.2874, -71.0646 42.286)"
+    )
+    factories.PotentialBreachFactory(
+        the_geom="SRID=4326;LINESTRING(-71.06452 42.287401, -71.064544 42.286)"
+    )
+    check = PotentialBreachInterdistanceCheck(
+        models.PotentialBreach.the_geom, min_distance=1.0
+    )
+    invalid = check.get_invalid(session)
+    assert len(invalid) == 0